+++ packages/core/package.json
@@ -27,6 +27,7 @@
     "access": "public"
   },
   "dependencies": {
+    "aesthetic-utils": "^2.0.0-alpha.0",
     "csstype": "^2.6.5",
     "extend": "^3.0.2",
     "rtl-css-js": "^1.13.0",
