--- packages/core/src/helpers/isRTL.ts
@@ -1,5 +1,0 @@
-import { Direction } from '../types';
-
-export default function isRTL(dir: Direction): boolean {
-  return dir === 'rtl';
-}
