+++ packages/core/src/testUtils.ts
@@ -3,9 +3,11 @@
 import CSS from 'csstype';
 import convertRTL from 'rtl-css-js';
 import Aesthetic from './Aesthetic';
-import getBaseFlushedStyles from './helpers/getFlushedStyles';
-import getStyleElements from './helpers/getStyleElements';
-import isObject from './helpers/isObject';
+import {
+  getFlushedStyles as getBaseFlushedStyles,
+  getStyleElements,
+  isObject,
+} from 'aesthetic-utils';
 import { FontFace, Direction } from './types';
 
 export { getStyleElements };
