+++ packages/utils/src/formatFontFace.ts
@@ -1,6 +1,11 @@
 import { Omit } from 'utility-types';
 import toArray from './toArray';
-import { FontFace } from '../types';
+
+interface FontFaceLike {
+  local?: string;
+  src?: string;
+  srcPaths: string[];
+}
 
 const FORMATS: { [ext: string]: string } = {
   '.eot': 'embedded-opentype',
@@ -12,9 +17,9 @@ const FORMATS: { [ext: string]: string } = {
   '.woff2': 'woff2',
 };
 
-export default function formatFontFace(
-  properties: Partial<FontFace>,
-): Omit<FontFace, 'local' | 'srcPaths'> {
+export default function formatFontFace<T extends FontFaceLike>(
+  properties: Partial<T>,
+): Omit<T, 'local' | 'srcPaths'> {
   const fontFace = { ...properties };
   const src: string[] = [];
