+++ packages/utils/src/hasQueryCondition.ts
