+++ packages/utils/src/isObject.ts
@@ -1,3 +1,3 @@
-export default function isObject(value: any): value is object {
+export default function isObject<T = object>(value: unknown): value is T {
   return value !== null && !Array.isArray(value) && typeof value === 'object';
 }
