+++ packages/utils/src/isRTL.ts
@@ -1,0 +1,3 @@
+export default function isRTL(dir: string): boolean {
+  return dir === 'rtl';
+}
