+++ packages/utils/src/stripClassPrefix.ts
